# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PetConnect API
  version: 1.0.0
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned after login or profile completion.
    SetupToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Setup token returned after OTP verification for onboarding.
  schemas:
...
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        favorites:
          type: array
          items:
            type: string
        location:
          $ref: "#/components/schemas/Location"
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
    UpdateAvatarInput:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl:
          type: string
          format: uri
    UpdateCoverInput:
      type: object
      required: [coverUrl]
      properties:
        coverUrl:
          type: string
          format: uri
...
    Location:
      type: object
      properties:
        city:
          type: string
          example: ""
        country:
          type: string
          example: IN
    PetMedicalRecord:
      type: object
      properties:
        title:
          type: string
          example: Vaccination Record
        documentUrl:
          type: string
          format: uri
        uploadedAt:
          type: string
          format: date-time
    PetHealthRecord:
      type: object
      properties:
        type:
          type: string
          enum: [vaccination, checkup, medication, tick_flea, surgery, dental, other]
        recordDetails:
          type: object
          properties:
            recordName:
              type: string
            batchLotNo:
              type: string
              nullable: true
            otherInfo:
              type: string
              nullable: true
            cost:
              type: string
              nullable: true
            date:
              type: string
              nullable: true
            nextDueDate:
              type: string
              nullable: true
            reminder:
              type: object
              properties:
                enabled:
                  type: boolean
                offset:
                  type: string
                  nullable: true
        veterinarian:
          type: object
          properties:
            designation:
              type: string
              nullable: true
            name:
              type: string
              nullable: true
            clinicName:
              type: string
              nullable: true
            licenseNo:
              type: string
              nullable: true
            contact:
              type: string
              nullable: true
        vitalSigns:
          type: object
          properties:
            weight:
              type: string
              nullable: true
            temperature:
              type: string
              nullable: true
            heartRate:
              type: string
              nullable: true
            respiratory:
              type: string
              nullable: true
            status:
              type: string
              enum: [normal, high, low]
              nullable: true
        observation:
          type: object
          properties:
            lookupObservations:
              type: array
              items:
                type: string
            clinicalNotes:
              type: string
              nullable: true
        attachments:
          type: array
          items:
            type: string
            format: uri
...
    Pet:
      type: object
      properties:
        id:
          type: string
          example: 64f1c3e2d2f2b70012a8b0aa
        owner:
          type: string
        name:
          type: string
        type:
          type: string
        species:
          type: string
        breed:
          type: string
          nullable: true
        age:
          type: number
          nullable: true
        weightLbs:
          type: number
          nullable: true
        gender:
          type: string
          enum: [male, female]
          nullable: true
        trained:
          type: boolean
          nullable: true
        vaccinated:
          type: boolean
          nullable: true
        neutered:
          type: boolean
          nullable: true
        personality:
          type: array
          items:
            type: string
        about:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          format: uri
          nullable: true
        photos:
          type: array
          items:
            type: string
            format: uri
        medicalRecords:
          type: array
          items:
            $ref: "#/components/schemas/PetMedicalRecord"
        healthRecords:
          type: array
          items:
            $ref: "#/components/schemas/PetHealthRecord"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: string
          example: 64f1c3e2d2f2b70012a8b0aa
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
        phone:
          type: string
          example: "9876543210"
        role:
          type: string
          example: user
        status:
          type: string
          enum: [active, deletion_request]
        bio:
          type: string
          example: ""
        username:
          type: string
          example: jane_doe
        avatarUrl:
          type: string
          format: uri
        coverUrl:
          type: string
          format: uri
        address:
          type: string
          example: ""
        location:
          $ref: "#/components/schemas/Location"
        favorites:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
          example: true
        isPhoneVerified:
          type: boolean
          example: false
        pets:
          type: array
          items:
            $ref: "#/components/schemas/Pet"
        profileCompleted:
          type: boolean
          example: true
        deletionRequestedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AdminUserDetails:
      type: object
      properties:
        id:
          type: string
          example: 64f1c3e2d2f2b70012a8b0aa
        name:
          type: string
          example: Jane Doe
        username:
          type: string
          example: jane_doe
        email:
          type: string
          format: email
          example: jane@example.com
        phone:
          type: string
          example: "9876543210"
        status:
          type: string
          enum: [active, deletion_request]
        deletionRequestedAt:
          type: string
          format: date-time
          nullable: true
        daysLeft:
          type: integer
          minimum: 0
          nullable: true
        isVerified:
          type: boolean
          example: true
        isPhoneVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DeletionRequestResponse:
      type: object
      properties:
        id:
          type: string
          example: 64f1c3e2d2f2b70012a8b0aa
        deletionRequestedAt:
          type: string
          format: date-time
...
paths:
...
  /admin/users/{id}:
    get:
      tags: [AdminUsers]
      summary: Get pet owner details
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/AdminUserDetails"
        "401":
          description: Unauthorized
        "404":
          description: User not found
    delete:
      tags: [AdminUsers]
      summary: Delete pet owner
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "401":
          description: Unauthorized
        "404":
          description: User not found
...
  /users/me:
    get:
      tags: [UserProfiles]
      summary: Get own profile
      security:
        - AccessToken: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      tags: [UserProfiles]
      summary: Update own profile
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/User"
        "400":
          description: Update failed
        "401":
          description: Unauthorized
  /users/me/password:
    patch:
      tags: [UserProfiles]
      summary: Change own password
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "400":
          description: Change failed
        "401":
          description: Unauthorized
  /users/me/avatar:
    patch:
      tags: [UserProfiles]
      summary: Update own avatar URL
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarInput"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/User"
        "400":
          description: Update failed
        "401":
          description: Unauthorized
  /users/me/cover:
    patch:
      tags: [UserProfiles]
      summary: Update own cover URL
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCoverInput"
      responses:
        "200":
          description: Cover updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/User"
        "400":
          description: Update failed
        "401":
          description: Unauthorized
  /users/me/delete-request:
    post:
      tags: [UserProfiles]
      summary: Request account deletion
      security:
        - AccessToken: []
      responses:
        "200":
          description: Deletion request received
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/DeletionRequestResponse"
        "400":
          description: Request failed
        "401":
          description: Unauthorized
  /users/me/delete-request/withdraw:
    post:
      tags: [UserProfiles]
      summary: Withdraw account deletion request
      security:
        - AccessToken: []
      responses:
        "200":
          description: Deletion request withdrawn
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/DeletionRequestResponse"
        "400":
          description: Withdraw failed
        "401":
          description: Unauthorized
  /users/{id}:
    get:
      tags: [UserProfiles]
      summary: Get user by id
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "404":
          description: User not found
...